+++ lib/kh2fm-randomizer/tsconfig.json
@@ -22,6 +22,6 @@
     "downlevelIteration": true,
 		"noFallthroughCasesInSwitch": true,
   },
-	"include": ["src", "src/index.test.ts", "src/index.ts"],
+	"include": ["src", "src/__tests__/index.test.ts", "src/index.ts"],
 	"exclude": ["node_modules", "src/**/*.test.ts", "dist"]
 }
